+++ packages/components/src/button/Button.tsx
@@ -189,7 +189,7 @@ export function Button({
       <AnimatedPressable
         {...props}
         accessibilityRole="button"
-        accessibilityState={{ disabled: isDisabled }}
+        aria-disabled={isDisabled}
         hitSlop={Platform.OS === 'web' ? undefined : 4}
         disabled={isDisabled}
         {...(isDisabled ? undefined : composedHandlers)}