+++ packages/components/src/radio/Radio.tsx
@@ -126,10 +126,8 @@ export function Radio({
     <AnimatedPressable
       {...props}
       accessibilityRole="radio"
-      accessibilityState={{
-        disabled: isDisabled,
-        checked: isSelected,
-      }}
+      aria-disabled={isDisabled}
+      aria-checked={isSelected}
       hitSlop={Platform.OS === 'web' ? undefined : 4}
       disabled={isDisabled}
       onPress={handlePress}